1. 修改music_score的解析逻辑 2. 将部分function修改为static 3. 增加枚举值Music::RollCall::ROLLCALL::ZERO 4. 处理部分warning

+++ MusicPlayer/src/player.cpp
@@ -44,7 +44,7 @@ void Player::Play()
             case Music::MusicalNote::MUSICALNOTE::EIGHTH_NOTE:
             case Music::MusicalNote::MUSICALNOTE::SIXTEENTH_NOTE:
             case Music::MusicalNote::MUSICALNOTE::THIRTY_SECOND_NOTE:
-                sleep = static_cast<float>(note) / m_rate;
+                sleep = static_cast<int>(static_cast<float>(note) / m_rate);
                 continue;
             case Music::MusicalNote::MUSICALNOTE::COMBO_NOTE_START:
             case Music::MusicalNote::MUSICALNOTE::COMBO_NOTE_STOP:
